MediaQueue rework.
6/1/2021 20:02

Even more major changes to MediaQueue.java
Finished the refactor for the queue system.
It can, theoretically, queue up various lists of Media subtypes, and said lists possess the source directory of the files, their extension, and their (collective) destination.
This means that it would be possible to queue up different lists of files that go to different destinations and have their own extensions.
It has been ported to a non-static implementation to allow for multiple instantiations.
It still has room for improvement as development progresses, but the core of it does not need to be changed for now.
General
Multiple files have had minor additions and changes to improve functionality.