Removal of GenericVideo.java
6/4/2021 10:05

Removal of GenericVideo.java
The initial goal for GenericVideo does not align with the direction of the program. Upon the implementation of user modification, the use case for generic video can be superseded by proper use of the Movie object.

Other minor changes have been made throughout the code to accommodate for this.
This also means that the only two Media subclasses are TV and Movie.

Development will proceed with these two in mind, but others, such as Music can be implemented at a later time.