SLF4J initial implementation
6/27/2021 10:52
General
Improved logging implementation with SLF4J.

Using SLF4J allows for dynamic logging implementations that do not tie up the entire codebase to a singular framework. This makes future-proofing the code easier and enhances interoperability.

When compiling, SLF4J must be added to the -classpath argument alongside the logging library of choice.
log4j was the one chosen for development.
More information can be found on SLF4J's site.
Logging is output to a log.out file in the working directory.

Added some logging functionality to various classes.

TV, Main, MetadataOps, and ConsoleEvent have had logging implemented or improved.
ConsoleEvent has been thoroughly reworked to accommodate these changes, and Logging (the Class) has been removed.
All classes will have logging implemented in the future.
Logs are no longer output on the console.

HandBrake (the Class) has been removed.

The implementation of this class was built on the idea of a single-minded and monolithic program. As a result, it clashes with the direction of the overall program.
The implementation of similar functionality is being considered but is not a priority; the stability and reliability of what's currently being worked on takes priority.