New branch: cli-abstraction
6/30/2021 15:41

ConsoleEvent has been removed from all major classes
ConsoleEvent will be used with the project, primarily within CLI applications.
This abstraction had reverberating effects throughout the project, all of which are notable.
MediaQueue
Now located in the media package.

The constructor now has four parameters:

The source directory of the files to search for.
The extension to filter with.
The destination directory.
This may be removed in favor of a postponed declaration.
The type to assign them as.
The arguments are checked before code execution for "validity" and will throw an IllegalArgumentException() if they are not valid.

Previous operations that prompted a program close through ConsoleEvent now log at the ERROR level.

MediaType
Renamed from MediaTypes.
The enum now features the following method.
public abstract Media instantiate (Path path);
This method must be overridden by every future MediaType.
The purpose of this change is to guarantee that every future MediaType possesses a way to instantiate its corresponding class, rather than having to maintain multiple areas of code whenever new ones are added.
It now has a logger to be primarily used whenever a given path does not correspond to a file.
Main
ConsoleEvent functionality from MediaQueue has been moved into a new method:
private static void userMediaQueueCLI ();