Safe file operations implemented.
7/5/2021 18:53
Implemented file checksum validation
This resides in yjohnson/Checksum and is hardcoded to use MD5.
This class has been designed to "provide support" to file operations by adding checksum validation.
Credit for large parts of this code go to Réal Gagnon @ Real's How-to.
Expanded functionality in Operations
New subclass: MediaIOWrapper.

It is meant to be a wrapper for Media objects that must undertake a move operation, providing a simpler way to pass along, track, and store Path information.
The method moveMedia() has been changed to copyMedia().

The method has been explicitly designed to do copy operations only.
This allows an intermediate step to occur between moving files: checksum validation.
The highlight of this method is the ability to use the Checksum class to verify that files are equivalent.
Because operations may take a long time, it is of utmost importance that the user's files are preserved throughout runtime. If the target directory is suddenly not reachable, full, or otherwise impaired, then the operation "fails" without compromising the user's files.

Functionality has been abstracted from organizedMove().

It now does a copy operation first and then deletes the source file, assuming the former succeeded.
This quite literally cripples per-file performance and doubles space complexity for the sake of safety, but I may look into developing an " unsafe" alternative in the future.
Source file deletion post-move is handled by removeSourceFilePostMove().
Media path generation has been moved to the Media subclasses with a method named generateCustomPathStructure().
Media & MediaQueue
Various method names have been changed to be more clear about their usage.

getMediaTitle() retrieves the media's proper title or name.

For the currently implemented classes, they imply:
TV - The series' name (not the episode's name)
Movie - The movie's name
This was added, in part, to improve MediaQueue.tallyNames().
generateCustomPathStructure() takes in a Path parameter and returns what the specific Media subclass would " want" their file structure to be under the given parent path.