Improvements to `Movie`
7/16/2021 15:39
Improvements to Movie
This class has been largely overhauled to adopt a similar construction to TV.

Release year parsing now utilizes regular expressions.

It is hardcoded to search for numbers in between 1800 and 2[century][decade]9, where the middle two are bounded by the current year.
This may pose a problem with names that contain a number such as this; user discretion is advised.
Logging has been added to various methods.

TV Abstraction
MetadataOps now houses removeUnwantedSpaces() and removeBrackets (originally found in TV).
removeBrackets can now remove leading or trailing brackets that do not have a counterpart, replacing them with an empty space character.