Queue and IO improvements
8/9/2021 14:47
Improvements to MediaQueue and MediaList
MediaQueue can now assign movies their own MediaList. * This is groundwork for possible (not guaranteed) changes, such as subtitle support and multi-part movies.
MediaList can now have its constructor invoked with a general Path that leads to a file or a directory.
This change is mostly to accommodate the above MediaQueue changes,
When given a file, the list will add the unique file to itself and return, essentially making a single-object list.
When given a directory, the constructor will function as it did before, searching for files within the path that meet the extension filter.
MediaList and MediaQueue are more verbose when logging.
PathFinder now uses an improved java.nio implementation
Credit for major parts of this code rewrite go to Yong Mook Kim (mkyong) and his tutorial @ mkyon.com.